Add help module, and many under the hood tweaks.
Fix naming override on modules.register.
Add automatic population of `help` and `full_help` on registered
  commands based on docstring.
Add hide, occludes, and occludable to modules.register kwargs.
- `hide` defaults to false, and keeps the command from being visible in
  command listings.
- `occludes` defaults to true, and will prevent other occludable
  commands from being activated after this one has.
- `occludable` defaults to true, and will prevent occluded messages from
  matching this command.

I'm not 100% happy with the occlusion system, but I need more time to
think on a better system and the help module relies on occlusion.